requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ADB Client
pure-python-adb>=0.3.0.dev0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0

# OCR
pytesseract>=0.3.10

# Utilities
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
